meson: fix test linking order
The static library we link against depends on symbols found in
the other link_withs. As meson passes --as-needed, we need to
ensure that the linker looks up all the undefined symbols in
the static library in the libraries that follow it on the command
line.

Part-of: <https://gitlab.freedesktop.org/gstreamer/meson-ports/ffmpeg/-/merge_requests/44>