libavfilter: Fix hang when linking with VS 2019 x64
If nasm is used to build an 'empty' assembly file, it still outputs an
object file for it but the MSVC linker seems to hang when linking to
it. This only happens for the v360 filter because the assembly helpers
are wrapped inside `%if HAVE_AVX2_EXTERNAL`.

Remove that if to fix the hang, and also because the if is wrong. The
functions should always be compiled and they will be plugged in at
runtime if AVX2 is found. That's exactly what `ff_v360_init_x86()`
inside `vf_v360_init.c` already does.

Reproduced with:
Microsoft (R) Incremental Linker Version 14.26.28806.0
NASM version 2.14.02 compiled on Dec 26 2018
Microsoft (R) C/C++ Optimizing Compiler Version 19.26.28806 for x64

Part-of: <https://gitlab.freedesktop.org/gstreamer/meson-ports/ffmpeg/-/merge_requests/46>